+++ app/controllers/blog_posts_controller.rb
@@ -1,7 +1,9 @@
 class BlogPostsController < ApplicationController
   def index
-    @blog_posts, @blog_posts_page, @blog_posts_total_pages =
-      paginate(BlogPost.order(created_at: :desc), per_page: 25)
+    data = BlogPosts::IndexData.call(paginator: method(:paginate))
+    @blog_posts = data.blog_posts
+    @blog_posts_page = data.blog_posts_page
+    @blog_posts_total_pages = data.blog_posts_total_pages
   end
 
   def show