+++ app/controllers/contact_controller.rb
@@ -1,20 +1,17 @@
 class ContactController < ApplicationController
   def index
-    path = Rails.root.join("config", "contact_info.yml")
-    @contact_info =
-      if File.exist?(path)
-        YAML.safe_load(File.read(path)) || {}
-      else
-        {}
-      end
+    data = Contact::IndexData.call
+    @contact_info = data.contact_info
   end
 
   def create
     return head :ok if params[:website].present?
 
-    ContactMailer
-      .contact_email(params[:name], params[:email], params[:message])
-      .deliver_later
+    Contact::SendMessage.call(
+      name: params[:name],
+      email: params[:email],
+      message: params[:message]
+    )
 
     redirect_to contact_path
   end