+++ app/services/about/index_data.rb
@@ -1,0 +1,13 @@
+module About
+  class IndexData
+    Result = Struct.new(:about_sections, keyword_init: true)
+
+    def self.call
+      new.call
+    end
+
+    def call
+      Result.new(about_sections: AboutSection.all)
+    end
+  end
+end